Working on setting up TinyMCE
TinyMCE is working and setup on the Create page. Currently exploring
options on which interface to use. Deciding on using the inline editing
mode in order to create a template and setting editable areas as needed.

Once configured, options to upload images need to be determined as well.
Multer is a good package to use when storing to MongoDB