Editor Page success posting to backend
Able to create a new article from the React Frontend and post the data
to the server.

Main image upload handled correctly. Need to change image upload handling
on the backend to an array to accomodate and images added to the article
body.

Need to add code to handle succesful submit and page redirect