Updated Client front end to display comments
CommentCard corrected typos.

ArticleSection separated useOutletContext variables since selectedArticle
is an Object and articleComments is an Array.

CommentSection map function was incorrectly written and is now corrected.

Article commentsResponseData code updated to provide an array instead of
an Object.

Comments useOutletContext variable was changed to accept Object.
map function was incorrectly written and is now corrected.